requirements/prod.txt
-r ./base.txt
gunicorn==23.0.0 # WSGI HTTP server for UNIX, used to serve Django applications.
redis
channels_redis
django-redis